requirements.txt
discord.py==2.1.0
requests==2.28.1
mysql-connector-python==8.0.31
pyspellchecker==0.7.0